Adding Features+Fixes to Script
Fixed issue where if the last signature was exactly a full signature long it would continue and error out.

Made every second page flipped so now everything prints as intended for generated signatures. (long-edge)

Now pdfs in the Input folder are automatically read and the whole script is ran for each pdf. After the numbered version and the signatures are generated for a pdf it is moved from the Input folder into its own folder in the Outputs.